refactor: extract slider parts to separate component
